exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exception.DatabaseException;
 import com.educandoweb.course.services.exception.ResourceNotFoundException;
 
 @Service
@@ -28,7 +30,14 @@ public class UserService {
 		return repository.save(obj);
 	}
 	public void delete (Long id) {
-		repository.deleteById(id);
+		 try {
+		        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+		        repository.deleteById(id);
+		    }catch (ResourceNotFoundException e){
+		        throw new ResourceNotFoundException(id);
+		    }catch (DataIntegrityViolationException e) {
+		    	throw new DatabaseException(e.getMessage());
+		    }
 	}
 	public User update (Long id, User obj) {
 		User entity = repository.getReferenceById(id);
